rename msg to msgs
